added Signup.js file in app.js
